% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_mean.R
\name{ct_mean}
\alias{ct_mean}
\title{Computes the arithmetic mean of the values of a vector.}
\usage{
ct_mean(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A number - the mean of the numbers in x.
}
\description{
Computes the arithmetic mean of the values of a vector.
}
\examples{
ct_mean(c(10,50,12,43))
}
