% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_meanCol.R
\name{ct_meanCol}
\alias{ct_meanCol}
\title{Computes the (arithmetic) mean value for each column in a matrix or a data frame.}
\usage{
ct_meanCol(x)
}
\arguments{
\item{x}{A matrix or a data frame.}
}
\value{
An n x 1 matrix with the mean value for each column in x.
The rownames are the colnames of x.
}
\description{
Computes the (arithmetic) mean value for each column in a matrix or a data frame.
}
\examples{
dataTest=cbind(c(10,50,12,43), c(100, 101, 102, 103), c(1, 1, 5,4))
colnames(dataTest)=c("Variable 1", "Variable 2", "Variable 3")
ct_meanCol(dataTest)
}
