% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_medianCol.R
\name{ct_medianCol}
\alias{ct_medianCol}
\title{Computes the median value for each column in a matrix or a data frame.}
\usage{
ct_medianCol(x)
}
\arguments{
\item{x}{A matrix or a data frame.}
}
\value{
An n x 1 matrix with the median value for each column in x.
The rownames are the colnames of x.
For each column, the function first sorts the values in ascending order. Then,
if the length of the column (=n) is divisible by 2, then the median will be
computed as the mean(C_n/2, C_n/2+1). Otherwise,
if the length of the column (=n) is NOT div by 2, then the median will be
computed as C_{n/2+1}=C_as.integer(n/2+1).
}
\description{
Computes the median value for each column in a matrix or a data frame.
}
\examples{
dataTest1=cbind(c(10,50,12,43), c(100, 101, 102, 103), c(1, 1, 5,4))
#note that length(dataTest1[,1]) is 4 and thus divisible by 2
colnames(dataTest1)=c("Variable 1", "Variable 2", "Variable 3")
ct_medianCol(dataTest1)

dataTest2=cbind(c(10,50,12,43,10), c(100, 101, 102, 103, 103), c(1, 1, 5,4, 7))
#note that length(dataTest2[,1]) is 5 and thus NOT divisible by 2
colnames(dataTest2)=c("Variable 1", "Variable 2", "Variable 3")
ct_medianCol(dataTest2)
}
