% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ct_mode.R
\name{ct_mode}
\alias{ct_mode}
\title{Finds the mode value(s) of a numeric vector.}
\usage{
ct_mode(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector - the mode value(s) of the numbers in x.
}
\description{
Finds the mode value(s) of a numeric vector.
}
\examples{
ct_mode(c(10,10,12,43))
ct_mode(c(10,43, 10,12,43))
}
